+++ Source/ACE.Entity/AceObjectPropertiesInstanceId.cs
@@ -3,7 +3,7 @@ using MySql.Data.MySqlClient;
 namespace ACE.Entity
 {
     [DbTable("ace_object_properties_iid")]
-    [DbGetList("ace_object_properties_iid", 22, "AceObjectId")]
+    [DbList("ace_object_properties_iid", "AceObjectId")]
     public class AceObjectPropertiesInstanceId
     {
         [DbField("AceObjectId", (int)MySqlDbType.UInt32, IsCriteria = true)]